fix requestId type : Int to Long

+++ spring_gateway/src/main/kotlin/com/example/spring_gateway/dto/response/FareRecommendResponse.kt
@@ -1,6 +1,6 @@
 package com.example.spring_gateway.dto.response
 
 class FareRecommendResponse(
-    var requestId:Int,
+    var requestId:Long,
     var recommendFare:Long
 )
